+++ example/native_test/run_native_tests.dart
@@ -1,7 +1,9 @@
 import 'streaming_error_test.dart' as streaming_error;
+import 'takepicture_test.dart' as takepicture;
 import 'transform_test.dart' as transform;
 
 void main() {
   streaming_error.main();
+  takepicture.main();
   transform.main();
 }